función para sacar el cpm del último año hecha y cambio en el modelo …

+++ FlaskApi/src/models/CPMModel.py
@@ -5,5 +5,5 @@ class CPM(db.Model):
     __tablename__ = 'CPM'
 
     cpmId = db.Column('cpmId', db.Integer, primary_key=True)
-    cpm = db.Column('cpm', db.Integer)
+    cpm = db.Column('cpm', db.Float)
     cpmDate = db.Column('cpmDate', db.DateTime, default=db.func.now())
